Lots of tweaks
